Fortigate-Netbox/requirements.txt
ipaddress
fortigate_api
pynetbox
python-dotenv